Improve pulp secrets code organization
[noissue]
