+++ controllers/repo_manager/api.go
@@ -18,8 +18,6 @@ package repo_manager
 
 import (
 	"context"
-	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/go-logr/logr"
@@ -246,278 +244,6 @@ func fileStoragePVC(resources controllers.FunctionResources) client.Object {
 	return pvc
 }
 
-// pulpServerSecret creates the pulp-server secret object which is used to
-// populate the /etc/pulp/settings.py config file
-func pulpServerSecret(resources controllers.FunctionResources) client.Object {
-
-	pulp := resources.Pulp
-	logger := resources.Logger
-	context := resources.Context
-	client := resources.Client
-
-	var dbHost, dbPort, dbUser, dbPass, dbName, dbSSLMode string
-	_, storageType := controllers.MultiStorageConfigured(pulp, "Pulp")
-
-	// if there is no external database configuration get the databaseconfig from pulp-postgres-configuration secret
-	if len(pulp.Spec.Database.ExternalDBSecret) == 0 {
-		postgresConfigurationSecret := pulp.Name + "-postgres-configuration"
-		resources.Logger.V(1).Info("Retrieving Postgres credentials from "+postgresConfigurationSecret+" secret", "Secret.Namespace", resources.Pulp.Namespace, "Secret.Name", resources.Pulp.Name)
-		pgCredentials, err := controllers.RetrieveSecretData(context, postgresConfigurationSecret, pulp.Namespace, true, client, "username", "password", "database", "port", "sslmode")
-		if err != nil {
-			logger.Error(err, "Secret Not Found!", "Secret.Namespace", pulp.Namespace, "Secret.Name", pulp.Name)
-		}
-		dbHost = pulp.Name + "-database-svc"
-		dbPort = pgCredentials["port"]
-		dbUser = pgCredentials["username"]
-		dbPass = pgCredentials["password"]
-		dbName = pgCredentials["database"]
-		dbSSLMode = pgCredentials["sslmode"]
-	} else {
-		resources.Logger.V(1).Info("Retrieving Postgres credentials from "+resources.Pulp.Spec.Database.ExternalDBSecret+" secret", "Secret.Namespace", resources.Pulp.Namespace, "Secret.Name", resources.Pulp.Name)
-		externalPostgresData := []string{"POSTGRES_HOST", "POSTGRES_PORT", "POSTGRES_USERNAME", "POSTGRES_PASSWORD", "POSTGRES_DB_NAME", "POSTGRES_SSLMODE"}
-		pgCredentials, err := controllers.RetrieveSecretData(context, pulp.Spec.Database.ExternalDBSecret, pulp.Namespace, true, client, externalPostgresData...)
-		if err != nil {
-			logger.Error(err, "Secret Not Found!", "Secret.Namespace", pulp.Namespace, "Secret.Name", pulp.Name)
-		}
-		dbHost = pgCredentials["POSTGRES_HOST"]
-		dbPort = pgCredentials["POSTGRES_PORT"]
-		dbUser = pgCredentials["POSTGRES_USERNAME"]
-		dbPass = pgCredentials["POSTGRES_PASSWORD"]
-		dbName = pgCredentials["POSTGRES_DB_NAME"]
-		dbSSLMode = pgCredentials["POSTGRES_SSLMODE"]
-	}
-
-	// Handling user facing URLs
-	rootUrl := getRootURL(resources)
-
-	// default settings.py configuration
-	var pulp_settings = controllers.DotNotEditMessage + `
-DB_ENCRYPTION_KEY = "/etc/pulp/keys/database_fields.symmetric.key"
-GALAXY_COLLECTION_SIGNING_SERVICE = "ansible-default"
-GALAXY_CONTAINER_SIGNING_SERVICE = "container-default"
-ANSIBLE_API_HOSTNAME = "` + rootUrl + `"
-ANSIBLE_CERTS_DIR = "/etc/pulp/keys/"
-CONTENT_ORIGIN = "` + rootUrl + `"
-DATABASES = {
-	'default': {
-		'HOST': '` + dbHost + `',
-		'ENGINE': 'django.db.backends.postgresql_psycopg2',
-		'NAME': '` + dbName + `',
-		'USER': '` + dbUser + `',
-		'PASSWORD': '` + dbPass + `',
-		'PORT': '` + dbPort + `',
-		'CONN_MAX_AGE': 0,
-		'OPTIONS': { 'sslmode': '` + dbSSLMode + `' },
-	}
-}
-GALAXY_FEATURE_FLAGS = {
-	'execution_environments': 'True',
-}
-PRIVATE_KEY_PATH = "/etc/pulp/keys/container_auth_private_key.pem"
-PUBLIC_KEY_PATH = "/etc/pulp/keys/container_auth_public_key.pem"
-STATIC_ROOT = "/var/lib/operator/static/"
-TOKEN_AUTH_DISABLED = False
-TOKEN_SIGNATURE_ALGORITHM = "ES256"
-`
-
-	pulp_settings = pulp_settings + fmt.Sprintln("API_ROOT = \"/pulp/\"")
-
-	// add cache settings
-	if pulp.Spec.Cache.Enabled {
-
-		var cacheHost, cachePort, cachePassword, cacheDB string
-
-		// if there is no ExternalCacheSecret defined, we should
-		// use the redis instance provided by the operator
-		if len(pulp.Spec.Cache.ExternalCacheSecret) == 0 {
-			if pulp.Spec.Cache.RedisPort == 0 {
-				cachePort = strconv.Itoa(6379)
-			} else {
-				cachePort = strconv.Itoa(pulp.Spec.Cache.RedisPort)
-			}
-			cacheHost = pulp.Name + "-redis-svc." + pulp.Namespace
-		} else {
-			// retrieve the connection data from ExternalCacheSecret secret
-			externalCacheData := []string{"REDIS_HOST", "REDIS_PORT", "REDIS_PASSWORD", "REDIS_DB"}
-			externalCacheConfig, _ := controllers.RetrieveSecretData(context, pulp.Spec.Cache.ExternalCacheSecret, pulp.Namespace, true, client, externalCacheData...)
-			cacheHost = externalCacheConfig["REDIS_HOST"]
-			cachePort = externalCacheConfig["REDIS_PORT"]
-			cachePassword = externalCacheConfig["REDIS_PASSWORD"]
-			cacheDB = externalCacheConfig["REDIS_DB"]
-		}
-
-		cacheSettings := `CACHE_ENABLED = True
-REDIS_HOST =  "` + cacheHost + `"
-REDIS_PORT =  "` + cachePort + `"
-REDIS_PASSWORD = "` + cachePassword + `"
-REDIS_DB = "` + cacheDB + `"
-`
-		pulp_settings = pulp_settings + cacheSettings
-	}
-
-	// if an Azure Blob is defined in Pulp CR we should add the
-	// credentials from azure secret into settings.py
-	if storageType[0] == controllers.AzureObjType {
-		resources.Logger.V(1).Info("Retrieving Azure data from " + resources.Pulp.Spec.ObjectStorageAzureSecret)
-		storageData, err := controllers.RetrieveSecretData(context, pulp.Spec.ObjectStorageAzureSecret, pulp.Namespace, true, client, "azure-account-name", "azure-account-key", "azure-container", "azure-container-path", "azure-connection-string")
-		if err != nil {
-			logger.Error(err, "Secret Not Found!", "Secret.Namespace", pulp.Namespace, "Secret.Name", pulp.Spec.ObjectStorageAzureSecret)
-			return &corev1.Secret{}
-		}
-		pulp_settings = pulp_settings + `AZURE_CONNECTION_STRING = '` + storageData["azure-connection-string"] + `'
-AZURE_LOCATION = '` + storageData["azure-container-path"] + `'
-AZURE_ACCOUNT_NAME = '` + storageData["azure-account-name"] + `'
-AZURE_ACCOUNT_KEY = '` + storageData["azure-account-key"] + `'
-AZURE_CONTAINER = '` + storageData["azure-container"] + `'
-AZURE_URL_EXPIRATION_SECS = 60
-AZURE_OVERWRITE_FILES = True
-DEFAULT_FILE_STORAGE = "storages.backends.azure_storage.AzureStorage"
-`
-	}
-
-	// if a S3 is defined in Pulp CR we should add the
-	// credentials from aws secret into settings.py
-	if storageType[0] == controllers.S3ObjType {
-		resources.Logger.V(1).Info("Retrieving S3 data from " + resources.Pulp.Spec.ObjectStorageS3Secret)
-		storageData, err := controllers.RetrieveSecretData(context, pulp.Spec.ObjectStorageS3Secret, pulp.Namespace, true, client, "s3-access-key-id", "s3-secret-access-key", "s3-bucket-name")
-		if err != nil {
-			logger.Error(err, "Secret Not Found!", "Secret.Namespace", pulp.Namespace, "Secret.Name", pulp.Spec.ObjectStorageS3Secret)
-			return &corev1.Secret{}
-		}
-
-		optionalKey, _ := controllers.RetrieveSecretData(resources.Context, resources.Pulp.Spec.ObjectStorageS3Secret, resources.Pulp.Namespace, false, client, "s3-endpoint", "s3-region")
-		if len(optionalKey["s3-endpoint"]) == 0 && len(optionalKey["s3-region"]) == 0 {
-			resources.Logger.Error(err, "Either s3-endpoint or s3-region needs to be specified", "Secret.Namespace", resources.Pulp.Namespace, "Secret.Name", resources.Pulp.Spec.ObjectStorageS3Secret)
-		}
-
-		if len(optionalKey["s3-endpoint"]) > 0 {
-			pulp_settings = pulp_settings + fmt.Sprintf("AWS_S3_ENDPOINT_URL = \"%v\"\n", optionalKey["s3-endpoint"])
-		}
-
-		if len(optionalKey["s3-region"]) > 0 {
-			pulp_settings = pulp_settings + fmt.Sprintf("AWS_S3_REGION_NAME = \"%v\"\n", optionalKey["s3-region"])
-		}
-
-		pulp_settings = pulp_settings + `AWS_ACCESS_KEY_ID = '` + storageData["s3-access-key-id"] + `'
-AWS_SECRET_ACCESS_KEY = '` + storageData["s3-secret-access-key"] + `'
-AWS_STORAGE_BUCKET_NAME = '` + storageData["s3-bucket-name"] + `'
-AWS_DEFAULT_ACL = "@none None"
-S3_USE_SIGV4 = True
-AWS_S3_SIGNATURE_VERSION = "s3v4"
-AWS_S3_ADDRESSING_STYLE = "path"
-DEFAULT_FILE_STORAGE = "storages.backends.s3boto3.S3Boto3Storage"
-MEDIA_ROOT = ""
-`
-	}
-
-	// configure settings.py with keycloak integration variables
-	if len(pulp.Spec.SSOSecret) > 0 {
-		ssoConfig(resources, &pulp_settings)
-	}
-
-	// configure TOKEN_SERVER based on ingress_type
-	tokenServer := "http://" + pulp.Name + "-api-svc." + pulp.Namespace + ".svc.cluster.local:24817/token/"
-	if isRoute(pulp) {
-		tokenServer = rootUrl + "/token/"
-	} else if isIngress(pulp) {
-		proto := "http"
-		if len(pulp.Spec.IngressTLSSecret) > 0 {
-			proto = "https"
-		}
-		hostname := pulp.Spec.IngressHost
-		tokenServer = proto + "://" + hostname + "/token/"
-	}
-	pulp_settings = pulp_settings + fmt.Sprintln("TOKEN_SERVER = \""+tokenServer+"\"")
-
-	if secretKey, err := loadDjangoKey(resources); err != nil {
-		return &corev1.Secret{}
-	} else {
-		pulp_settings = pulp_settings + fmt.Sprintln("SECRET_KEY = \""+secretKey+"\"")
-	}
-
-	// add custom settings to the secret
-	pulp_settings = addCustomPulpSettings(pulp, pulp_settings)
-
-	sec := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pulp.Name + "-server",
-			Namespace: pulp.Namespace,
-		},
-		StringData: map[string]string{
-			"settings.py": pulp_settings,
-		},
-	}
-
-	// Set Pulp instance as the owner and controller
-	ctrl.SetControllerReference(pulp, sec, resources.Scheme)
-	return sec
-}
-
-// pulp-db-fields-encryption secret
-func pulpDBFieldsEncryptionSecret(resources controllers.FunctionResources) client.Object {
-	pulp := resources.Pulp
-	sec := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pulp.Name + "-db-fields-encryption",
-			Namespace: pulp.Namespace,
-		},
-		StringData: map[string]string{
-			"database_fields.symmetric.key": createFernetKey(),
-		},
-	}
-	return sec
-}
-
-// pulp-admin-password
-func pulpAdminPasswordSecret(resources controllers.FunctionResources) client.Object {
-
-	pulp := resources.Pulp
-	sec := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pulp.Spec.AdminPasswordSecret,
-			Namespace: pulp.Namespace,
-		},
-		StringData: map[string]string{
-			"password": createPwd(32),
-		},
-	}
-	ctrl.SetControllerReference(pulp, sec, resources.Scheme)
-
-	return sec
-}
-
-// pulpDjangoKeySecret defines the Secret with the pulp-secret-key
-func pulpDjangoKeySecret(resources controllers.FunctionResources) client.Object {
-	pulp := resources.Pulp
-	sec := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pulp.Spec.PulpSecretKey,
-			Namespace: pulp.Namespace,
-		},
-		StringData: map[string]string{
-			"secret_key": djangoKey(),
-		},
-	}
-	ctrl.SetControllerReference(pulp, sec, resources.Scheme)
-	return sec
-}
-
-// pulp-container-auth
-func pulpContainerAuth(resources controllers.FunctionResources) client.Object {
-	pulp := resources.Pulp
-	privKey, pubKey := genTokenAuthKey()
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pulp.Spec.ContainerTokenSecret,
-			Namespace: pulp.Namespace,
-		},
-		StringData: map[string]string{
-			"container_auth_private_key.pem": privKey,
-			"container_auth_public_key.pem":  pubKey,
-		},
-	}
-}
-
 // serviceForAPI returns a service object for pulp-api
 func serviceForAPI(resources controllers.FunctionResources) client.Object {
 	pulp := resources.Pulp
@@ -585,17 +311,4 @@ func serviceAPISpec(name, namespace, deployment_type string) corev1.ServiceSpec 
 func storageClassProvided(pulp *repomanagerpulpprojectorgv1beta2.Pulp) bool {
 	_, storageType := controllers.MultiStorageConfigured(pulp, "Pulp")
 	return storageType[0] == controllers.SCNameType
-}
-
-// loadDjangoKey retrieves the django SECRET_KEY from the k8s Secret
-func loadDjangoKey(resources controllers.FunctionResources) (string, error) {
-	pulp := resources.Pulp
-	djangoKey := pulp.Spec.PulpSecretKey
-	resources.Logger.V(1).Info("Retrieving the data from "+pulp.Spec.PulpSecretKey+" Secret", "Secret.Namespace", pulp.Namespace, "Secret.Name", djangoKey)
-	secretKey, err := controllers.RetrieveSecretData(resources.Context, djangoKey, pulp.Namespace, true, resources.Client, "secret_key")
-	if err != nil {
-		resources.Logger.Error(err, "Secret Not Found!", "Secret.Namespace", pulp.Namespace, "Secret.Name", djangoKey)
-		return "", err
-	}
-	return secretKey["secret_key"], nil
 }
